+++ src/main/java/io/xlorey/fluxloader/client/core/Core.java
@@ -23,10 +23,9 @@ public class Core {
     public static boolean isPluginsExecuted = false;
 
     /**
-     * Initializing the loader
-     * @exception Exception in cases of unsuccessful core initialization
+     * Pre-initialization of the core. Loads before all game initializations.
      */
-    public static void init() throws Exception {
+    public static void preInit() {
         Logger.printCredits();
 
         Logger.print("FluxLoader Core initialization for the client...");
@@ -35,7 +34,14 @@ public class Core {
         EventManager.subscribe(new OnMainMenuEnterHandler(), EventPriority.HIGHEST);
         EventManager.subscribe(new OnPostTickRenderThreadHandler(), EventPriority.HIGHEST);
         EventManager.subscribe(new OnPostUIDrawHandler(), EventPriority.HIGHEST);
+    }
 
+    /**
+     * Initializing the loader. Called after loading mods and basic initializations,
+     * namely at the moment after the Translator is initialized
+     * @exception Exception in cases of unsuccessful core initialization
+     */
+    public static void init() throws Exception {
         PluginManager.loadPlugins(true);
 
         WidgetManager.init();